pyproject.toml
[project]
name = "deepmosa-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asciitree==0.3.3",
    "astor~=0.8.1",
    "astroid==3.3.8",
    "black~=25.1.0",
    "bytecode~=0.16.1",
    "colorama~=0.4.6",
    "graphviz~=0.17",
    "grappa~=1.0.1",
    "jellyfish==1.1.3",
    "matplotlib~=3.10.0",
    "networkx==3.4.2",
    "openai~=1.97.1",
    "pandas==2.2.3",
    "pipreqs~=0.5.0",
    "pycg==0.0.6",
    "pyexcel-ods3~=0.6.1",
    "pytest~=8.3.4",
    "python-dotenv~=1.1.0",
    "regex==2023.3.23",
    "requests==2.32.3",
    "ruff==0.0.260",
    "seaborn~=0.13.2",
    "setuptools>=80.9.0",
    "simple-parsing~=0.1.7",
    "transformers~=4.49.0",
    "typed-ast==1.5.4",
    "typeshed-client~=2.2.0",
    "typing-inspect~=0.9.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "isort~=5.12.0",
    "pytest-asyncio~=1.1.0",
]

[project.scripts]
pynguin = "pynguin.cli:main"

[tool.setuptools.packages.find]
include = ["pynguin*", "vendor*", "app*"]
